Moving around some of the helper functions
